Rename propagation modules
