+++ instrumentation/akka-actor-2.5/src/test/groovy/AkkaExecutorInstrumentationTest.groovy
