+++ instrumentation/akka-actor-2.5/src/test/scala/AkkaAsyncChild.java
