+++ instrumentation/scala-executors/src/test/groovy/ScalaInstrumentationTest.groovy
