+++ settings.gradle
@@ -53,7 +53,7 @@ include ':utils:test-utils'
 // smoke tests
 include ':smoke-tests'
 
-include ':instrumentation:akka-context-propagation-2.5'
+include ':instrumentation:akka-actor-2.5'
 include ':instrumentation:akka-http-10.0'
 include ':instrumentation:apache-camel-2.20'
 include ':instrumentation:apache-httpasyncclient-4.0'
@@ -125,7 +125,7 @@ include ':instrumentation:jsp-2.3'
 include ':instrumentation:kafka-clients-0.11'
 include ':instrumentation:kafka-streams-0.11'
 include ':instrumentation:khttp-0.1'
-include ':instrumentation:kotlin-coroutines'
+include ':instrumentation:kotlinx-coroutines'
 include ':instrumentation:kubernetes-client-7.0'
 include ':instrumentation:lettuce:lettuce-common'
 include ':instrumentation:lettuce:lettuce-4.0'
@@ -169,7 +169,7 @@ include ':instrumentation:redisson-3.0'
 include ':instrumentation:rmi'
 include ':instrumentation:rxjava-1.0'
 include ':instrumentation:rxjava-1.0'
-include ':instrumentation:scala-concurrent'
+include ':instrumentation:scala-executors'
 include ':instrumentation:servlet:glassfish-testing'
 include ':instrumentation:servlet:servlet-common'
 include ':instrumentation:servlet:servlet-2.2'
